requirements.txt
streamlit
pandas
plotly
fpdf
google-genai
